C++: construct defaults for record fields
A default-constructed GenericDatum has type AVRO_NULL. When used as a
default for a field, this typically signifies no default value. But it
is actually meaningful for unions: it means the default is null. This
can make the difference between a program crashing because of a missing
field or not. This commit adds the default-constructed datum as the
default for all record fields.